event-management-system copy/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
bcrypt
python-multipart
python-dotenv
email-validator
typing-extensions 